feature/blog : 블로그 기능 관련 api 및 기능 추가

+++ src/main/java/org/example/story/domain/portfolio/service/PortfolioCommentService.java
@@ -7,7 +7,6 @@ import org.example.story.domain.portfolio.record.request.PortfolioCommentRequest
 import org.example.story.domain.portfolio.record.response.PortfolioCommentListResponse;
 import org.example.story.domain.portfolio.record.response.PortfolioCommentResponse;
 import org.example.story.domain.portfolio.repository.PortfolioCommentRepository;
-import org.example.story.domain.portfolio.repository.PortfolioLikeRepository;
 import org.example.story.domain.portfolio.repository.PortfolioRepository;
 import org.example.story.domain.user.entity.UserJpaEntity;
 import org.example.story.domain.user.repository.UserRepository;
@@ -53,6 +52,10 @@ public class PortfolioCommentService {
     public void deleteComment(Long userId, Long portfolioId, Long commentId) {
         PortfolioCommentJpaEntity comment = portfolioCommentRepository.findByPortfolioIdAndId(portfolioId,commentId)
                 .orElseThrow(() -> new ExpectedException(HttpStatus.NOT_FOUND, "존재하지 않는 포트폴리오입니다."));
+
+        if (!comment.getUser().getId().equals(userId)) {
+            throw new ExpectedException(HttpStatus.FORBIDDEN, "댓글을 삭제할 권한이 없습니다.");
+        }
         portfolioCommentRepository.delete(comment);
     }
